requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
audioop-lts==0.2.2
click==8.3.1
colorama==0.4.6
discord.py==2.6.4
fastapi==0.127.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
yarl==1.22.0
